Move logic to their own methods
